src/lambdas/train/requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
boto3==1.28.37
python-dotenv==0.21.1